requirements.txt
coreapi
coreschema
Django
djangorestframework
drf-yasg
packaging
pytz
Markdown
Pygments
django-filter
django-guardian
whitenoise
gunicorn
django-cors-headers
django-createsuperuser
xlwt
reportlab
Pillow
django-jenkins
python-docx
XlsxWriter
python-pptx
ezodf 
lxml
gTTS
pyunpack
djangorestframework_simplejwt
qrcode
snappy
python-snappy
requests-kerberos
pywinrm
pymongo[snappy,gssapi,srv,tls,zstd]
